Fix for the undefined context
The issue was that inside of Squire#remove there is a call to completely
destroy the current Squire.js context that RequireJS was managing. This
context was the context created during the constructor of the Squire
instance.

The solution implemented in this patch is to rebuild a new RequireJS
context, after deleting the initial context. This way the RequireJS
context is fresh, as the remove call would indicate it should be, while
also still being defined for subsequent calls.

Additionally, there should no longer be an edge case that leads to the
context being undefined in the first place, as was the initial symptom
of the issue.